requirements.txt
allure-pytest==2.15.0
flake8==7.3.0
pytest-rerunfailures==12.0
pytest==7.4.3
PyYAML==6.0.1
selenium==4.15.2